Rewrite download_file
